Removed filter by discipline and disabled deletion of scheduled interview

+++ apps/api/src/interview-slot/interview-slot.service.ts
@@ -4,7 +4,6 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InterviewStatus } from '@prisma/client';
-import { Discipline } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
 
 import { CreateInterviewSlotDto } from './dto/createInterviewSlot.dto';
@@ -27,74 +26,18 @@ export class InterviewSlotService {
     return interviewSlots;
   }
 
-  async getByInterviewers(interviewerNames: string[]) {
-    const interviewSlots = await this.prisma.interviewSlot.findMany({
-      where: {
-        interviewers: {
-          some: {
-            interviewer: {
-              name: {
-                in: interviewerNames,
-              },
-            },
-          },
-        },
-      },
-      include: {
-        interviewers: {
-          include: {
-            interviewer: true,
-            interviewSlot: true,
-          },
-        },
-      },
-    });
-    return interviewSlots;
-  }
-
-  async getByDisciplines(disciplines: string[]) {
-    const mappedDisciplines: Discipline[] = disciplines.map((str) => {
-      switch (str) {
-        case 'Development':
-          return Discipline.Development;
-        case 'Design':
-          return Discipline.Design;
-        case 'Multimedia':
-          return Discipline.Multimedia;
-        case 'Marketing':
-          return Discipline.Marketing;
-        default:
-          return Discipline.Development;
-      }
-    });
-
-    const interviewSlots = await this.prisma.interviewSlot.findMany({
-      where: {
-        interviewers: {
-          some: {
-            interviewer: {
-              disciplines: {
-                hasSome: mappedDisciplines,
-              },
-            },
-          },
-        },
-      },
-      include: {
-        interviewers: {
-          include: {
-            interviewer: true,
-            interviewSlot: true,
-          },
-        },
-      },
-    });
-    return interviewSlots;
-  }
-
   async deleteInterviewSlot(id: string) {
-    return await this.prisma.interviewSlot.delete({
+    const interviewToDelete = await this.prisma.interviewSlot.findUnique({
       where: { id },
+      include: {
+        intern: true,
+      },
+    });
+
+    if (interviewToDelete.internId) return;
+
+    return await this.prisma.interviewSlot.delete({
+      where: { id: interviewToDelete.id },
     });
   }
